### 2019-01-24 #1199 @whedon
[PRE REVIEW]: straditize: Digitizing stratigraphic diagrams
**Submitting author:** @Chilipp (<a href="http://orcid.org/0000-0001-6171-7716">Philipp S. Sommer</a>)
**Repository:** <a href="https://github.com/Chilipp/straditize" target ="_blank">https://github.com/Chilipp/straditize</a>
**Version:** v0.1.0
