### 2019-01-25 #1199 @Chilipp
I'd say ideally someone with a background in GUI development and/or digitisation. Maybe @ixjlyons or @sauln?
