### 2019-01-25 #1199 @sauln
I'd love to help, but don't have any experience in digitization.
