### 2019-01-26 #1199 @ixjlyons
I'm also not very knowledgable in this area, but my understanding of the review process leads me to believe I'm qualified to do the review and I'd be happy to do it. @yochannah what level of familiarity with the field do you expect a reviewer to have?
