### 2019-01-26 0822c13
Added menu-windows.json to conda-recipe
