+++ conda-recipe/meta.yaml
@@ -9,7 +9,7 @@ source:
   path: ../
 
 build:
-  number: 0
+  number: 1
   script: python -m pip install . --no-deps --ignore-installed -vvv
   entry_points:
     - straditize = straditize.__main__:main
