### 2019-01-28 75a148e
minor bug fix to ensure slicing with integers
