### 2019-01-28 #1199 @sgrieve
@yochannah I'd be delighted to review. This looks really cool!
