### 2019-01-28 6c5eb50
fixed a bug for bar reader
If a column doesn't contain any bars, the software should not fail.
