### 2019-01-28 #1216 @Chilipp
> [ 👉 Check article proof 📄 👈 ](https://github.com/openjournals/joss-papers/blob/joss.01216/joss.01216/10.21105.joss.01216.pdf)

seems all right to me. Thanks at @sgrieve and @ixjlyons for being willing to review this package. Looking forward to your feedbacks!
