### 2019-01-28 b2f8b2c
do not replace occurences_value in xaxis-tranlation
