### 2019-02-05 Chilipp/straditize#2 @ixjlyons
Conda package not found on Linux
Hello, here from the JOSS review [openjournals/joss-reviews#1216](https://github.com/openjournals/joss-reviews/issues/1216)

#### Problem description

I ran the following:

```
$ conda create -n straditize
$ conda activate straditize
$ conda install -c chilipp straditize
Collecting package metadata: done
Solving environment: failed

PackagesNotFoundError: The following packages are not available from current channels:

  - straditize
<clip>
```

I'm not all that experienced with conda package management, but looking at [this page](https://anaconda.org/Chilipp/straditize/files), it looks like only packages with the `master` label are available for Linux, so I figured out that the install from git version worked:

```
conda install -c chilipp/label/master straditize
```

#### Environment information

```conda info -a``` (the ```<details>``` environment doesn't seem to support formatting):

```
     active environment : straditize
    active env location : /home/kenny/miniconda3/envs/straditize
            shell level : 1
       user config file : /home/kenny/.condarc
 populated config files : /home/kenny/.condarc
          conda version : 4.6.2
    conda-build version : not installed
         python version : 3.6.7.final.0
       base environment : /home/kenny/miniconda3  (writable)
           channel URLs : https://conda.anaconda.org/conda-forge/linux-64
                          https://conda.anaconda.org/conda-forge/noarch
                          https://repo.anaconda.com/pkgs/main/linux-64
                          https://repo.anaconda.com/pkgs/main/noarch
                          https://repo.anaconda.com/pkgs/free/linux-64
                          https://repo.anaconda.com/pkgs/free/noarch
                          https://repo.anaconda.com/pkgs/r/linux-64
                          https://repo.anaconda.com/pkgs/r/noarch
          package cache : /home/kenny/miniconda3/pkgs
                          /home/kenny/.conda/pkgs
       envs directories : /home/kenny/miniconda3/envs
                          /home/kenny/.conda/envs
               platform : linux-64
             user-agent : conda/4.6.2 requests/2.21.0 CPython/3.6.7 Linux/4.20.6-arch1-1-ARCH arch/rolling glibc/2.28
                UID:GID : 1000:1000
             netrc file : None
           offline mode : False

# conda environments:
#
base                     /home/kenny/miniconda3
straditize            *  /home/kenny/miniconda3/envs/straditize

sys.version: 3.6.7 | packaged by conda-forge | (defau...
sys.prefix: /home/kenny/miniconda3
sys.executable: /home/kenny/miniconda3/bin/python
conda location: /home/kenny/miniconda3/lib/python3.6/site-packages/conda
conda-build: None
conda-env: /home/kenny/miniconda3/bin/conda-env
user site dirs:

CIO_TEST: <not set>
CONDA_DEFAULT_ENV: straditize
CONDA_PREFIX: /home/kenny/miniconda3/envs/straditize
CONDA_PROMPT_MODIFIER: (straditize)
CONDA_PYTHON_EXE: /home/kenny/miniconda3/bin/python
CONDA_ROOT: /home/kenny/miniconda3
CONDA_SHLVL: 1
PATH: /home/kenny/miniconda3/envs/straditize/bin:/home/kenny/.pyenv/shims:/home/kenny/miniconda3/condabin:/home/kenny/.pyenv/bin:/home/kenny/.pyenv/bin:/usr/local/sbin:/usr/local/bin:/usr/bin:/usr/bin/site_perl:/usr/bin/vendor_perl:/usr/bin/core_perl:/home/kenny/bin:/home/kenny/.local/bin:/usr/bin:/sbin:/home/kenny/bin:/home/kenny/.local/bin
REQUESTS_CA_BUNDLE: <not set>
SSL_CERT_FILE: <not set>
WINDOWPATH: 1
```
