### 2019-02-05 Chilipp/straditize#3 @ixjlyons
A couple usability notes
Hello, here from the JOSS review [openjournals/joss-reviews#1216](https://github.com/openjournals/joss-reviews/issues/1216)

In this issue I'll just describe a couple stumbling points I came across in my testing of straditize. I think in both cases, a quick note in the docs at the right point would have been helpful.

It took me quite a while to figure out that you can disable the matplotlib pan and zoom controls by clicking on the corresponding button again. This made the instructions confusing in the beginner tutorial because I couldn't figure out how to select and drag the diagram boundary limits. I think the place to mention this would be [here](https://straditize.readthedocs.io/en/latest/tutorial/beginner/beginner-tutorial-select-data.html).

It also took some time to figure out that I need to click-and-hold on the wand toolbar button to get the submenu to show up like on [this page](https://straditize.readthedocs.io/en/latest/tutorial/beginner/beginner-tutorial-clean-image.html). I don't think I've seen a user interface like that outside mobile applications -- is that the default behavior in Qt and is it the same for other platforms?
