### 2019-02-05 Chilipp/straditize#4 @ixjlyons
A few broken links in the documentation
Hello, here from the JOSS review [openjournals/joss-reviews#1216](https://github.com/openjournals/joss-reviews/issues/1216)

I just came across a few broken links in the docs in my review.

- Link to the code of conduct [here](https://straditize.readthedocs.io/en/latest/contribute.html#code-of-conduct)
- Link to the issue templates [here](https://straditize.readthedocs.io/en/latest/contribute.html#reporting-bugs) and [here](https://straditize.readthedocs.io/en/latest/contribute.html#how-do-i-submit-a-good-bug-report)
- Link to pull request template [here](https://straditize.readthedocs.io/en/latest/.github/PULL_REQUEST_TEMPLATE.md)
