### 2019-02-05 Chilipp/straditize#2 @Chilipp
Dear @ixjlyons,

you are completely right, thanks for submitting and my apologies! This was because the CI build of the master branch [#153](https://travis-ci.org/Chilipp/straditize/builds/483932906) happened shortly after the CI build of the release [#152](https://travis-ci.org/Chilipp/straditize/builds/483932899).
The install instructions will however anyway to use the `conda-forge` channel instead. For this, however, https://github.com/conda-forge/staged-recipes/pull/7658 first needs to be resolved and merged.

I updated the labels in the [Chilipp conda channel](https://anaconda.org/Chilipp/straditize/files) such that you can install (the build 0) with `conda install -c chilipp straditize`

Does this resolve your issue?
