### 2019-02-05 #1216 @ixjlyons
Ok, my review is complete. Nice work @Chilipp!

Overall, straditize appears to be in excellent shape. Aside from a couple installation issues (filed against the target repository), all of the review criteria appear to me to be met. Beyond those, the code looks clean and well-written. I'm not all that familiar with digitization, but adding to the kinds of plots that can be "reverse-engineered" definitely seems worthwhile. The paper conveys that sentiment well and summarizes the package adequately. The documentation looks good standalone, but it shines inside the actual GUI and the integration with the tutorials is excellent.

Once the two installation issues I ran into are addressed, I recommend this submission be accepted.
