### 2019-02-05 e2e6614
Added PyQt5 as dependency
