### 2019-02-05 aab569a
added note disabling on pan and zoom tools
this commit fixes the missing documentation on using matplotlibs pan and zoom tools together with the marks (see https://github.com/Chilipp/straditize/issues/3)
