### 2019-02-05 30b3cdd
added docs on selection tool menus
This commit adds documentation of how to open the context menus of the selection tools (see https://github.com/Chilipp/straditize/issues/3)
