### 2019-02-05 8a4b67e
Skip PyQt5 depencency on CONDA_BUILD
