### 2019-02-05 051100d
fixed broken links for contribution guide
see https://github.com/Chilipp/straditize/issues/4
