### 2019-02-05 28d2f96
Reduce verbosity of tests
