### 2019-02-06 #1216 @Chilipp
Dear @ixjlyons, thank you very much for your time and your helpful feedback! I already implemented the suggested changes will merge them soon.
