### 2019-02-06 Chilipp/straditize#2 @ixjlyons
Yep, seems to be working now. Thanks.
