### 2019-02-07 1ee381a
fix to libgfortran=3 on osx
see https://github.com/conda-forge/numpy-feedstock/issues/126
