### 2019-02-08 Chilipp/straditize#8 @sgrieve
Tests abort on MacOS
#### Problem description

Hello from the JOSS review [openjournals/joss-reviews#1216](https://github.com/openjournals/joss-reviews/issues/1216).

I have run into some trouble getting the tests to pass on my machine. Straditize itself seems to be working fine however.

When I run `py.test` or `python setup.py test` I get the following output:

```
[stuart:~/straditize] [strad] master 21m1s ± py.test
platform darwin -- Python 3.6.3, pytest-4.2.0, py-1.7.0, pluggy-0.8.1
rootdir: /Users/stuart/straditize, inifile:
collected 179 items

tests/test_binary.py .......                                             [  3%]
tests/test_straditizer.py ..                                             [  5%]
tests/widgets/test_axes_translations.py ..                               [  6%]
tests/widgets/test_beginner.py ............                              [ 12%]
tests/widgets/test_colnames.py sss..s                                    [ 16%]
tests/widgets/test_command_line.py Abort trap: 6
```

#### Environment information

I am running MacOS Sierra 10.12.6 with Python 3.6.3. I have created a clean Python 3.6.3 virtual environment in which I have run `python setup.py install`, `pip install pyqt5` and `pip install pytest`.

<summary>psyplot -aV</summary>

# Paste the output of the command ``psyplot -aV`` (ran from the command line)

```
psyplot:
  requirements:
    matplotlib: 3.0.2
    numpy: 1.16.1
    pandas: 0.24.1
    python: 3.6.3 |Anaconda, Inc.| (default, Oct  6 2017, 12:04:38)  [GCC 4.2.1 Compatible
      Clang 4.0.1 (tags/RELEASE_401/final)]
    xarray: 0.11.3
  version: 1.2.0
psyplot_gui:
  requirements:
    pyqt: 5.11.3
    qt: 5.11.2
    qtconsole: 4.4.3
  version: 1.2.2
```

<summary>straditize -V</summary>

# Paste the output of the command ``straditize -V`` (ran from the command line)
```
0.1.0
```
