### 2019-02-08 Chilipp/straditize#8 @Chilipp
Dear @sgrieve, thanks for reporting this! May I ask you how you installed straditize (`pip install .`, `python setup.py install`, `conda install`, etc.) and to provide the output of `pip list` (or `conda list` if you used `conda`)
