### 2019-02-08 Chilipp/straditize#8 @sgrieve
Hi @Chilipp I installed using `python setup.py install`

and the output of pip list is:

```
Package                  Version
------------------------ ----------
alabaster                0.7.12
appnope                  0.1.0
atomicwrites             1.3.0
attrs                    18.2.0
Babel                    2.6.0
backcall                 0.1.0
certifi                  2018.11.29
cftime                   1.0.3.4
chardet                  3.0.4
cloudpickle              0.7.0
cycler                   0.10.0
dask                     1.1.1
decorator                4.3.2
docrep                   0.2.5
docutils                 0.14
et-xmlfile               1.0.1
fasteners                0.14.1
funcargparse             0.2.0
idna                     2.8
imagesize                1.1.0
ipykernel                5.1.0
ipython                  7.2.0
ipython-genutils         0.2.0
jdcal                    1.4
jedi                     0.13.2
Jinja2                   2.10
jupyter-client           5.2.4
jupyter-core             4.4.0
kiwisolver               1.0.1
MarkupSafe               1.1.0
matplotlib               3.0.2
monotonic                1.5
more-itertools           5.0.0
netCDF4                  1.4.2
networkx                 2.2
numpy                    1.16.1
openpyxl                 2.6.0
packaging                19.0
pandas                   0.24.1
parso                    0.3.3
pexpect                  4.6.0
pickleshare              0.7.5
Pillow                   5.4.1
pip                      19.0.1
pluggy                   0.8.1
prompt-toolkit           2.0.8
psyplot                  1.2.0
psyplot-gui              1.2.2
ptyprocess               0.6.0
py                       1.7.0
Pygments                 2.3.1
pyparsing                2.3.1
PyQt5                    5.11.3
PyQt5-sip                4.19.13
pytest                   4.2.0
python-dateutil          2.8.0
pytz                     2018.9
PyWavelets               1.0.1
PyYAML                   4.2b4
pyzmq                    17.1.2
qtconsole                4.4.3
requests                 2.21.0
scikit-image             0.14.2
scipy                    1.2.0
setuptools               40.8.0
six                      1.12.0
snowballstemmer          1.2.1
Sphinx                   1.8.4
sphinx-rtd-theme         0.4.2
sphinxcontrib-websupport 1.1.0
straditize               0.1.0
toolz                    0.9.0
tornado                  6.0a1
traitlets                4.3.2
urllib3                  1.24.1
wcwidth                  0.1.7
wheel                    0.32.3
xarray                   0.11.3
```
