### 2019-02-08 #1216 @sgrieve
Hi @Chilipp, @yochannah Below is my review, as noted below I have also opened an issue in the straditize repo regarding the tests.

Firstly I would like to thank the author for this contribution. I have been very impressed by the quality of the software presented here, as well as it's potential utility to a wide range of users across disciplines. I am happy to recommend publication, pending the resolution of the issues outlined below.

# Installation

I am performing this review on MacOS Sierra 10.12.6, and found no problems with the conda install as outlined in the documentation. However, one thing I did note is that the quick install instructions on the GitHub readme does not recommend installing in a separate virtual environment. I feel that installing into a virtual environment as the default quickstart install option is probably better than potentially causing problems in someones base environment.

I had the same problem as @ixjlyons when trying to install via pip and from source, that I needed to manually install `PyQt`. It seems that you have already addressed this, but I just wanted to highlight that it was an issue on my environment as well.

# Functionality

Having worked through the tutorials I am happy to confirm that straditize works as described. More generally, I believe that this will be a significant tool for scientists in a range of disciplines, and the way that the software has been designed suggests to me that it will be straightforward to add more diagram types to straditize in future based on community need.

# Documentation

I am very impressed by the level and quality of the documentation, in particular the integration of the documentation and tutorials into the GUI. Aside from the broken links mentioned elsewhere in this review, I noted a typo on the first line of this section: https://straditize.readthedocs.io/en/latest/installing.html#running-the-tests

`We us` should be `We use`.

# Community guidelines

I just wanted to put a note here that your contribution/community guidelines are excellent. I will return to these the next time I have to write some for one of my projects. As above, I noted some broken links, but these have all been addressed in PR 7.

# Tests

I have not been able to get the tests to pass on my machine. I have opened an issue to outline this further: https://github.com/Chilipp/straditize/issues/8

It may be useful to add some detail of what a user can expect when running these tests, for example the number of tests that will be run, what does a successful test run look like, are any tests skipped, the approximate duration of the tests.

# Paper

The paper is very well written and clearly outlines the scientific problem the author is trying to address, as well as how straditize specifically addresses these issues. The format of the paper follows the general JOSS style, and I believe that all of the relevant literature and materials have been correctly cited.

In conclusion, I have thoroughly enjoyed the experience of reviewing this paper and would like to commend the author for their diligent work to address a pressing concern in our field. I am hoping to introduce this tool to some of my students in the future and look forward to using it in my own research. :rocket:
