### 2019-02-09 #1216 @yochannah
Wow, thanks @ixjlyons and @sgrieve for your thorough reviews, and @Chilipp for acting on the feedback so swiftly!

@sgrieve @Chilipp let me know when you two have managed to resolve the testing issue and are ready to go ahead! 👍
