### 2019-02-09 #1216 @Chilipp
Dear @yochannah, @sgrieve and @ixjlyons,

thank you very much for your nice words, helpful feedbacks and the reported issues. I apologize for the inconveniences with pip and also with the tests @sgrieve. Unfortunately I have no good methodology to test such a GUI-based application. The unittest framework requires creating and closing lot's of widgets which produces quite some overhead (this especially causes problems on the CI services with very limited RAM, such as OSX on Travis). I would like to address these problems first before merging the suggested changes. But this should be done by Monday.
