### 2019-02-10 e989b18
improved efficiency and removed leaks
This commit adds a leaking report for pytest runs and decreases several memory leaks during the test runs
