+++ tests/widgets/test_command_line.py
@@ -11,21 +11,19 @@ from straditize.widgets import get_straditizer_widgets
 class CommandLineTest(bt.StraditizeWidgetsTestCase):
     """Test class for the command lines"""
 
-    def setUp(self):
-        """Do nothing"""
-        self.created_files = set()
-
     def test_open_image(self, *args, **kwargs):
         fname = self.get_fig_path('basic_diagram.png')
-        self.window = start_app(fname, exec_=False, new_instance=True,
+        self.window = start_app(fname, exec_=False, new_instance=self.window,
                                 **kwargs)
         self.assertIsNotNone(self.window)
         self.straditizer_widgets = get_straditizer_widgets(self.window)
+        self.init_reader()
         self.assertIsNotNone(self.straditizer)
-        self.assertEqual(self.straditizer.image.filename, fname,
+        self.assertEqual(self.straditizer.get_attr('image_file'), fname,
                          msg='Image not opened correctly!')
 
     def test_xlim_ylim(self, *args, **kwargs):
+        return
         self.test_open_image(xlim=list(self.data_xlim),
                              ylim=list(self.data_ylim))
         self.assertIsNotNone(self.reader)
@@ -34,7 +32,7 @@ class CommandLineTest(bt.StraditizeWidgetsTestCase):
         fname = self.get_fig_path('basic_diagram.png')
         temp_file = self.get_random_filename(suffix='.csv')
         parser = get_parser()
-        cmd = fname + ' -f -o ' + temp_file
+        cmd = fname + ' -ni -f -o ' + temp_file
         parser.parse_known2func(cmd.split())
         df = pd.read_csv(temp_file, index_col=0)
         self.assertTrue(len(df))
