### 2019-02-10 b3d4d71
add psutil to conda-recipes test requirements
