+++ conda-recipe/meta.yaml
@@ -54,9 +54,15 @@ test:
     - pytest -v --cov=straditize
       --ignore=tests/widgets/test_selection_toolbar.py
       --ignore=tests/widgets/test_samples_table.py
+      --ignore=tests/widgets/test_beginner.py
+      --ignore=tests/widgets/test_hoya_del_castillo.py
     - pytest -v --cov=straditize --cov-append
       tests/widgets/test_selection_toolbar.py
       tests/widgets/test_samples_table.py
+    - pytest -v --cov=straditize --cov-append
+      tests/widgets/test_beginner.py
+    - pytest -v --cov=straditize --cov-append
+      tests/widgets/test_hoya_del_castillo.py
     - codecov
 
 
