### 2019-02-11 195c36e
Add travis_wait for conda build
