### 2019-02-11 6d88b04
skip memory intensive tests on OSX on Travis
