+++ .travis.yml
@@ -93,7 +93,7 @@ install:
 script:
     - unset_secure
     # install conda
-    - travis_wait 30 conda build conda-recipe --python $PYTHON_VERSION
+    - travis_wait 30 conda build conda-recipe --python $PYTHON_VERSION --clobber-file ci/recipe_clobber_${TRAVIS_OS_NAME}.yaml
     # export the pw to make it available for the deploy
     - reset_secure
     # deploy here (don't use travis deploy because it requires homebrew)
