### 2019-02-11 cda94dc
Improved conda installation and testing instructions
in response to https://github.com/openjournals/joss-reviews/issues/1216#issuecomment-461775141
