### 2019-02-11 a561cba
include running tests with setuptools
