+++ .travis.yml
@@ -3,7 +3,6 @@ sudo: false # use container based build
 
 env:
     global:
-        - PYTHON_VERSION=3.6
         - secure: "m3hFOyDrNPNT+UfC3jbn9QY4ZYK2DBgZDmF5Nf1jy7+gpERLWLy+BuG+bO7WifliZe6BUD5DNG9qkHewaRnYKiLJhyMK3yaS1iLSV6NLjfxG9bsnPzC8FCF/aB8EgYpA5pH2Wk6NEM/e+3QjwKHWMgia2HLeZ1AXmRQuzNzGUb0Ayad+EogZetygmVDiTiDbMOJ1hNQb9racYeGRgyyjY+wrQ11yjXqhwDRlAR6sHo98VDwPav+ecjyc7YY+y7bmoJ5TbpMg/VCY3wSgZPSdLEwsBAuTejI/U5bAzqo5lYy+tzqYLgWvRTHPwJ1aRfWdRd9LYJY0IfpPBYeZhuHi0K/yME0Kx4Dm+zZBpKrUF/mLrd6dfGLd3nc8xp6NtuEuaSKVcumtrRImbQOvnWx9fZxLKc0qwa0mZhWxZuWgNPpXBCcb6bgHAutK7To+eiCUJ6Swe/YZIHNK/t6Ag+cUisBY3AJpz6GVZIkgTY7GZCGCnYBWO/4USR840HBy5oDH1LUPxTiCcpFvWyOTC8w6n3mwoGQi2g6hZ11HtDE3DFYan1hntYk9aU2nE+705e9Hi/rRNvWpzzS/o1fqmEq6Pn3vA3RHUrKzpEl4coKpf0Wa9ZGRj3v267X+eY0r+uijigZC3eW2gG0x2pfEnNEsahHzAynr1Pxi3FJnys8a+nw="
 
 matrix:
@@ -12,6 +11,10 @@ matrix:
             - PYTHON_VERSION=3.7
           os: linux
         - env:
+            - PYTHON_VERSION=3.7
+              USE_PIP=true
+          os: linux
+        - env:
             - PYTHON_VERSION=3.6
           os: linux
         - env:
@@ -20,6 +23,12 @@ matrix:
         - env:
             - PYTHON_VERSION=3.6
           os: osx
+        - env:
+            - PYTHON_VERSION=3.7
+              USE_PIP=true
+          os: osx
+    allow_failures:
+        - env: USE_PIP=true
 
 before_install:
     # Remove homebrew.
@@ -89,14 +98,29 @@ install:
     - export PYTHONWARNINGS='ignore:mode:DeprecationWarning:docutils.io:245'
     - conda info -a
     - conda list
+    - if [[ $USE_PIP == true ]]; then
+          TEST_CMD="
+              python setup.py test -a '
+                  --ignore=tests/widgets/test_selection_toolbar.py
+                  --ignore=tests/widgets/test_samples_table.py
+                  --ignore=tests/widgets/test_beginner.py
+                  --ignore=tests/widgets/test_hoya_del_castillo.py' &&
+              python setup.py test -a 'tests/widgets/test_selection_toolbar.py
+                  tests/widgets/test_samples_table.py' &&
+              python setup.py test -a 'tests/widgets/test_beginner.py' &&
+              python setup.py test -a 'tests/widgets/test_hoya_del_castillo.py'";
+      else
+          TEST_CMD="conda build conda-recipe --python $PYTHON_VERSION --clobber-file ci/recipe_clobber_${TRAVIS_OS_NAME}.yaml";
+      fi
+    - echo $TEST_CMD
     - reset_secure
 script:
     - unset_secure
     # install conda
[…]
