### 2019-02-11 2fe33ba
force psyplot-gui>=1.2.3
