### 2019-02-11 d4c1860
install via pip when USE_PIP
