### 2019-02-11 2dd18f0
use pip for installation instead of setuptools
