### 2019-02-11 2e7256b
fixed rst syntax [skip ci]
