### 2019-02-11 Chilipp/straditize#1 @Chilipp
Dear @ixjlyons. I'm finally coming back to you. I implemented the changes in the `pip-install-fix` branch (see https://github.com/Chilipp/straditize/pull/6/). Could you please check out this branch (``git checkout pip-install-fix``) and run your installation commands again?
