### 2019-02-11 Chilipp/straditize#8 @Chilipp
Dear @sgrieve! Thanks again for reporting this issue. Unfortunately PyQt always aborts the test run and therefore makes it difficult to debug. Nevertheless, I think I figured it out by running `py.test -s` to not capture the stdout. With this, I got the error message

```
Traceback (most recent call last):
  File "/Users/psommer/Documents/myplots-scripts/psyplot-gui/psyplot_gui/console.py", line 314, in eventFilter
    return super().eventFilter(*args, **kwargs)
  File "/Users/psommer/miniconda3/envs/test2/lib/python3.7/site-packages/qtconsole/console_widget.py", line 407, in eventFilter
    self._trigger_is_complete_callback()
  File "/Users/psommer/miniconda3/envs/test2/lib/python3.7/site-packages/qtconsole/console_widget.py", line 578, in _trigger_is_complete_callback
    if self._is_complete_msg_id is not None:
AttributeError: 'ConsoleWidget' object has no attribute '_is_complete_msg_id'
```
 This is one of the difficulties with unittests and PyQt. The `ConsoleWidget` this error comes from has actually already been closed and is a remainder from the previous `TestCase`. Nevertheless PyQt still calls the `eventFilter` method of this widget which produces the error above.

Unfortunately the created widget is not garbage collected although the corresponding widget has already been closed. Usually this is not a problem in straditize (you would not close the console and open a new one) but within the unittests I create and close lot's of widgets.

However, I fixed this in the psyplot-gui package (https://github.com/Chilipp/psyplot-gui/commit/23656d1e78d59fbc7f4da60a2c19e5a5e5448ada) and made a new release (1.2.3) that is available on [PyPi](https://pypi.org/project/psyplot-gui/) and will soon be available from the conda-forge channel (see https://github.com/conda-forge/psyplot-gui-feedstock/pull/9) and is already available on the [Chilipp conda channel](https://anaconda.org/Chilipp/psyplot-gui).

Could you please checkout the `pip-install-fix` branch of straditize (see https://anaconda.org/Chilipp/psyplot-gui) and check that the tests run through? Please also consider the changed instructions on [how to run the tests](https://straditize.readthedocs.io/en/pip-install-fix/installing.html#running-the-tests) and [how to install it from source](https://straditize.readthedocs.io/en/pip-install-fix/installing.html#installation-from-source) in the docs. The problem with `pyqt5` through pip should be solved now and is also addressed in the CI build.

Thank you very much for your engagement! You should not, but if you still encounter the previous problem (or any new problem), please include the output of `py.test -s` in your report.
