### 2019-02-11 Chilipp/straditize#4 @Chilipp
Dear @ixjlyons, thanks again for reporting this issue! I addressed it in https://github.com/Chilipp/straditize/pull/7#issuecomment-462485712, could you please confirm that it fixes this issue before I merge the PR?
