### 2019-02-11 Chilipp/straditize#3 @Chilipp
Dear @ixjlyons, thanks again for reporting your experiences here! It is always difficult to figure such usability issues out as the developer of the software. I addressed the proposed changes in the [tutorial-doc-fix](https://github.com/Chilipp/straditize/tree/tutorial-doc-fix). Could you please tell me if the proposed changes resolve your issues (see https://github.com/Chilipp/straditize/pull/5#issuecomment-462488675).
