### 2019-02-12 Chilipp/straditize#8 @sgrieve
Hi @Chilipp The tests all now pass for me and the changes to the documentation are excellent. Great work!
