### 2019-02-12 Chilipp/straditize#8 @Chilipp
Thank you 👍  And thanks for the quick response! I will close this issue then when https://github.com/Chilipp/straditize/pull/6 is merged
