### 2019-02-12 5522234
minor doc changes [skip ci]
in response to https://github.com/Chilipp/straditize/pull/6#discussion_r256035234 and https://github.com/Chilipp/straditize/pull/6#discussion_r256035592
