### 2019-02-12 0e9f8af
minor doc changes [skip ci]
in response to https://github.com/Chilipp/straditize/pull/6#discussion_r256035976
