### 2019-02-12 Chilipp/straditize#1 @ixjlyons
I'm not sure if you get an email notification when a PR is reviewed, but this and the other two issues I opened look like they're resolved once the corresponding PRs are merged.
