### 2019-02-12 Chilipp/straditize#1 @Chilipp
> I'm not sure if you get an email notification when a PR is reviewed

I got a notification. Thank you :)
