### 2019-02-12 9fc92ce
Revert "fix to libgfortran=3 on osx"
This reverts commit 07f16963126d3db892213b0ffec35b257bf9b011
based on
https://github.com/Chilipp/straditize/commit/1ee381a47ec0a0e3fc9dc534102f65acf6a166fb#r32285992
