### 2019-02-13 Chilipp/straditize#4 @Chilipp
Closed by https://github.com/Chilipp/straditize/pull/7
