### 2019-02-13 Chilipp/straditize#1 @Chilipp
Closed by https://github.com/Chilipp/straditize/pull/6
