### 2019-02-13 Chilipp/straditize#8 @Chilipp
Closed byhttps://github.com/Chilipp/straditize/pull/6
