### 2019-02-13 Chilipp/straditize#3 @Chilipp
Closed by https://github.com/Chilipp/straditize/pull/5
