### 2019-02-13 #1216 @Chilipp
Dear @yochannah, @sgrieve and @ixjlyons,

thanks again for all the input and your fast responses! All reported issues (https://github.com/Chilipp/straditize/issues/1, https://github.com/Chilipp/straditize/issues/2, https://github.com/Chilipp/straditize/issues/3, https://github.com/Chilipp/straditize/issues/4 and https://github.com/Chilipp/straditize/issues/8) are closed and from my side, I am ready to make a new release with a fresh DOI. Is this okay for you?
