### 2019-02-13 #1216 @yochannah
Okay, I've given it a finalish proofread and the text looks great 💯 . One strange thing I noted is that the figure appears in the middle of the reference section though - @arfon any idea what is up?
