+++ CHANGELOG.rst
@@ -1,0 +1,20 @@
+v0.1.1
+======
+
+This release has been approved by the Journal of Open Source Software
+(JOSS) in https://github.com/openjournals/joss-reviews/issues/1216
+
+Added
+-----
+* Changelog
+
+Changed
+-------
+* Thanks to the feedbacks from `@ixjlyons <https://github.com/ixjlyons>`__ and
+  `@sgrieve <https://github.com/sgrieve`__ the installation instructions and
+  some other documentation instructions have been improved (see the issues
+  https://github.com/Chilipp/straditize/issues/1,
+  https://github.com/Chilipp/straditize/issues/2,
+  https://github.com/Chilipp/straditize/issues/3,
+  https://github.com/Chilipp/straditize/issues/4 and
+  https://github.com/Chilipp/straditize/issues/8)
