### 2019-02-13 #1216 @yochannah
@sgrieve and @ixjlyons, y'all each have one checkbox left unticked. If you're happy to accept, can you check those boxes and make a comment once you're done?

@Chilipp Once those last two checkboxes are complete, please do make an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive!  It's probably worth giving the manuscript a final proofread if you feel you need to, as well. 👍

Thanks all, it's been really positive watching this all come together! 🥇
