### 2019-02-13 #1216 @arfon
> @arfon any idea what is up?

@yochannah - this is standard LaTeX image layout issues. If the authors make this change the image placement is better: https://github.com/Chilipp/straditize/pull/9
