+++ straditize/version.py
@@ -17,4 +17,4 @@ GNU General Public License for more details.
 
 You should have received a copy of the GNU General Public License
 along with this program. If not, see <https://www.gnu.org/licenses/>."""
-__version__ = '0.1.0'
+__version__ = '0.1.1'
