### 2019-02-14 #1216 @ixjlyons
@yochannah I checked the last box assuming @Chilipp will make a release on PyPI soon based on the recent fixes so `pip install straditize` Just Works&trade; per the installation instructions. Should a new release be made before getting a DOI?
