### 2019-02-14 #1216 @Chilipp
All right @yochannah, I made a new version 0.1.1 that is available on PyPi and conda and created a release on zenodo:

[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.2565036.svg)](https://doi.org/10.5281/zenodo.2565036)

> It's probably worth giving the manuscript a final proofread if you feel you need to, as well

Also for the manuscript everything looks all right.
