### 2019-02-17 #1216 @Chilipp
> @Chilipp okay, looking good - one last thing before we accept - can you make sure the Zenodo archive title is the same as the paper title? Thanks!

All right, done

https://zenodo.org/record/2565036#.XGmbPNF7mM8
