+++ conda-recipe/meta.yaml
@@ -9,19 +9,14 @@ source:
   path: ../
 
 build:
-  number: 1
+  number: 2
   script: python -m pip install . --no-deps --ignore-installed -vvv
+  skip: true  # [py == 27]
   entry_points:
     - straditize = straditize.__main__:main
 
-app:
-  entry: straditize
-  icon: logo.png
-  summary: Stratigraphic Diagram Digitization software
-  type: desk
-
 requirements:
-  build:
+  host:
     - python
     - pip
   run:
@@ -29,9 +24,9 @@ requirements:
     - psyplot-gui >=1.2.2
     - psyplot >=1.2.0
     - netcdf4
-    - pyqt >=5.7.0
+    - pyqt >=5
     - scipy
-    - numpy >=1.14  # [win]
+    - numpy >=1.12  # [win]
     - scikit-image
     - openpyxl
     - sphinx !=1.7.5
@@ -65,6 +60,11 @@ test:
       tests/widgets/test_hoya_del_castillo.py
     - codecov
 
+app:
+  entry: straditize
+  icon: logo.png
+  summary: Stratigraphic Diagram Digitization software
+  type: desk
 
 about:
   home: https://github.com/Chilipp/straditize
