### 2019-02-17 04e5d86
added installation instructions for conda-forge
