### 2019-02-17 #1216 @yochannah
@openjournals/joss-eics - we're ready to accept this one!! 🎉 🏆

@Chilipp Thanks for your great work on this package and fantastic collaboration with the reviewers.

@sgrieve @ixjlyons Thank you for your fantastic and in-depth reviews.
